fix: permitir decimales en el valor de la cuota en UF
El campo amount estaba tipado como Integer en el modelo y como int en
los schemas, asi que cualquier cuota en UF con decimales era rechazada
antes de llegar a la base de datos.

- amount pasa a Numeric(12, 2) / Decimal, manteniendo la validacion gt=0
- el payload a BUK convierte a float en UF y a int en pesos, porque
  Decimal no es serializable a JSON
- el input del formulario usa step/min 0.01 cuando la moneda es UF
- migracion 013 para cambiar el tipo de la columna existente

La migracion se aplica a mano en la VPS:
psql -d rh_cramer -f backend/migrations/013_creditos_amount_decimal.sql

+++ backend/app/models/creditos.py
@@ -17,7 +17,7 @@ class Credito(Base):
     tipo_prestamo = Column(String(50), nullable=False, default="Préstamo Emergencia")
     start_date = Column(Date, nullable=False)
     moneda = Column(String(10), nullable=False, default="peso")
-    amount = Column(Integer, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     cuota_actual = Column(Integer, nullable=False, default=1)
     duracion = Column(Integer, nullable=False)
     monto_original = Column(Numeric(12, 2), nullable=True)
